Adding a MRVI class wrapper to torchmrvi and jaxmrvi and having all functionality working, using a backend parameter
