+++ src/scvi/external/mrvi/__init__.py
@@ -1,4 +1,3 @@
 from ._model import MRVI
-from ._module import MRVAE
 
-__all__ = ["MRVI", "MRVAE"]
+__all__ = ["MRVI"]
