+++ src/scvi/external/mrvi_jax/__init__.py
@@ -1,0 +1,4 @@
+from ._model import JaxMRVI
+from ._module import JaxMRVAE
+
+__all__ = ["JaxMRVI", "JaxMRVAE"]
