+++ tests/external/mrvi_torch/test_torchmrvi_model.py
@@ -7,7 +7,7 @@ import numpy as np
 import pytest
 
 from scvi.data import synthetic_iid
-from scvi.external import TorchMRVI
+from scvi.external import MRVI
 
 if TYPE_CHECKING:
     from typing import Any
@@ -37,14 +37,14 @@ def adata():
 
 @pytest.fixture(scope="session")
 def model(adata: AnnData):
-    TorchMRVI.setup_anndata(adata, sample_key="sample_str", batch_key="batch")
-    model = TorchMRVI(adata)
+    MRVI.setup_anndata(adata, sample_key="sample_str", batch_key="batch", backend="torch")
+    model = MRVI(adata, backend="torch")
     model.train(max_steps=1, train_size=0.5)
 
     return model
 
 
-def test_torchmrvi(model: TorchMRVI, adata: AnnData, save_path: str):
+def test_torchMRVI(model: MRVI, adata: AnnData, save_path: str):
     model.get_local_sample_distances()
     model.get_local_sample_distances(normalize_distances=True)
     model.get_latent_representation(give_z=False)
@@ -52,7 +52,7 @@ def test_torchmrvi(model: TorchMRVI, adata: AnnData, save_path: str):
 
     model_path = os.path.join(save_path, "mrvi_model")
     model.save(model_path, save_anndata=False, overwrite=True)
-    model = TorchMRVI.load(model_path, adata=adata)
+    model = MRVI.load(model_path, adata=adata, backend="torch")
 
 
 @pytest.mark.optional
@@ -101,7 +101,7 @@ def test_torchmrvi(model: TorchMRVI, adata: AnnData, save_path: str):
         ),
     ],
 )
-def test_torchmrvi_de(model: TorchMRVI, setup_kwargs: dict[str, Any], de_kwargs: dict[str, Any]):
+def test_torchMRVI_de(model: MRVI, setup_kwargs: dict[str, Any], de_kwargs: dict[str, Any]):
     for de_kwarg in de_kwargs:
         model.differential_expression(**de_kwarg)
 
@@ -121,7 +121,7 @@ def test_torchmrvi_de(model: TorchMRVI, setup_kwargs: dict[str, Any], de_kwargs:
         {"sample_cov_keys": ["meta1_cat", "batch"], "compute_log_enrichment": True},
     ],
 )
-def test_torchmrvi_da(model, sample_key, da_kwargs):
+def test_torchMRVI_da(model, sample_key, da_kwargs):
     model.differential_abundance(**da_kwargs)
 
 
@@ -155,37 +155,39 @@ def test_torchmrvi_da(model, sample_key, da_kwargs):
         },
     ],
 )
-def test_torchmrvi_model_kwargs(adata: AnnData, model_kwargs: dict[str, Any], save_path: str):
-    TorchMRVI.setup_anndata(
+def test_torchMRVI_model_kwargs(adata: AnnData, model_kwargs: dict[str, Any], save_path: str):
+    MRVI.setup_anndata(
         adata,
         sample_key="sample_str",
         batch_key="batch",
+        backend="torch",
     )
-    model = TorchMRVI(adata, n_latent=10, scale_observations=True, **model_kwargs)
+    model = MRVI(adata, n_latent=10, scale_observations=True, backend="torch", **model_kwargs)
     model.train(max_steps=1, train_size=0.5)
 
     model_path = os.path.join(save_path, "mrvi_model")
     model.save(model_path, save_anndata=False, overwrite=True)
-    model = TorchMRVI.load(model_path, adata=adata)
+    model = MRVI.load(model_path, adata=adata, backend="torch")
 
 
-def test_torchmrvi_sample_subset(model: TorchMRVI, adata: AnnData, save_path: str):
+def test_torchMRVI_sample_subset(model: MRVI, adata: AnnData, save_path: str):
     sample_cov_keys = ["meta1_cat", "meta2", "cont_cov"]
     sample_subset = [chr(i + ord("a")) for i in range(8)]
     model.differential_expression(sample_cov_keys=sample_cov_keys, sample_subset=sample_subset)
 
     model_path = os.path.join(save_path, "mrvi_model")
     model.save(model_path, save_anndata=False, overwrite=True)
-    model = TorchMRVI.load(model_path, adata=adata)
+    model = MRVI.load(model_path, adata=adata, backend="torch")
 
 
-def test_torchmrvi_shrink_u(adata: AnnData, save_path: str):
-    TorchMRVI.setup_anndata(
+def test_torchMRVI_shrink_u(adata: AnnData, save_path: str):
+    MRVI.setup_anndata(
         adata,
         sample_key="sample_str",
         batch_key="batch",
+        backend="torch",
     )
-    model = TorchMRVI(adata, n_latent=10, n_latent_u=5)
+    model = MRVI(adata, n_latent=10, n_latent_u=5, backend="torch")
     model.train(max_steps=1, train_size=0.5)
     model.get_local_sample_distances()
 
@@ -196,7 +198,7 @@ def test_torchmrvi_shrink_u(adata: AnnData, save_path: str):
 
     model_path = os.path.join(save_path, "mrvi_model")
     model.save(model_path, save_anndata=False, overwrite=True)
-    model = TorchMRVI.load(model_path, adata=adata)
+    model = MRVI.load(model_path, adata=adata, backend="torch")
 
 
 @pytest.fixture
@@ -213,13 +215,14 @@ def adata_stratifications():
     return adata
 
 
-def test_torchmrvi_stratifications(adata_stratifications: AnnData, save_path: str):
-    TorchMRVI.setup_anndata(
+def test_torchMRVI_stratifications(adata_stratifications: AnnData, save_path: str):
+    MRVI.setup_anndata(
         adata_stratifications,
         sample_key="sample_str",
         batch_key="batch",
+        backend="torch",
     )
-    model = TorchMRVI(adata_stratifications, n_latent=10)
+    model = MRVI(adata_stratifications, n_latent=10, backend="torch")
     model.train(max_steps=1, train_size=0.5)
 
     dists = model.get_local_sample_distances(groupby=["labels", "label_2"])
@@ -234,4 +237,4 @@ def test_torchmrvi_stratifications(adata_stratifications: AnnData, save_path: st
 
     model_path = os.path.join(save_path, "mrvi_model")
     model.save(model_path, save_anndata=False, overwrite=True)
-    model = TorchMRVI.load(model_path, adata=adata_stratifications)
+    model = MRVI.load(model_path, adata=adata_stratifications, backend="torch")
